+++ AllAboutGames.Data/Models/GameGenre.cs
@@ -7,18 +7,16 @@ namespace AllAboutGames.Data.Models
     public class GameGenre
     {
         [Key]
-        public int GameGenreID { get; set; }
+        public long GameGenreID { get; set; }
 
         [Required]
         [ForeignKey(nameof(Game))]
-        public int GameID { get; set; }
-
+        public long GameID { get; set; }
         public virtual Game Game { get; set; }
 
         [Required]
         [ForeignKey(nameof(Genre))]
-        public int GenreID { get; set; }
-
+        public long GenreID { get; set; }
         public virtual Genre Genre { get; set; }
     }
 }
