+++ AllAboutGames.Data/Models/GamePlatform.cs
@@ -7,18 +7,16 @@ namespace AllAboutGames.Data.Models
     public class GamePlatform
     {
         [Key]
-        public int GamePlatformID { get; set; }
+        public long GamePlatformID { get; set; }
 
         [Required]
         [ForeignKey(nameof(Game))]
-        public int GameID { get; set; }
-
+        public long GameID { get; set; }
         public virtual Game Game { get; set; }
 
         [Required]
         [ForeignKey(nameof(Platform))]
-        public int PlatformID { get; set; }
-
+        public long PlatformID { get; set; }
         public virtual Platform Platform { get; set; }
     }
 }
