+++ AllAboutGames.Data/Models/Rating.cs
@@ -11,14 +11,12 @@ namespace AllAboutGames.Data.Models
 
         [Required]
         [ForeignKey(nameof(Game))]
-        public int GameID { get; set; }
-
+        public long GameID { get; set; }
         public virtual Game Game { get; set; }
 
         [Required]
         [ForeignKey(nameof(User))]
         public long UserID { get; set; }
-
         public virtual ApplicationUser User { get; set; }
 
         [Required(ErrorMessage = "Rating value is required.")]
